Minor relocating files.
