+++ backend/src/main/java/com/gymlogger/backend/controller/auth/AuthenticationController.java
@@ -1,6 +1,5 @@
 package com.gymlogger.backend.controller.auth;
 
-import com.gymlogger.backend.service.AuthenticationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
